# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Health OS Backend API
  version: 1.0.0
servers:
  - url: https://staging.api.healthos.app
  - url: http://localhost:8080
paths:
  /v1/auth/register:
    post:
      tags: [AUTH]
      summary: Registro comun para Mobile y Web
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/auth/login:
    post:
      tags: [AUTH]
      summary: Login Mobile con JSON tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileLoginResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/auth/web/login:
    post:
      tags: [AUTH]
      summary: Login Web con HttpOnly cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso con Set-Cookie access_token y csrf_token
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebLoginResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/auth/refresh:
    post:
      tags: [AUTH]
      summary: Rotacion on-use de Refresh Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Nuevos tokens JWT; el refresh anterior queda revocado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileLoginResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/sync/measurements:
    post:
      tags: [SYNC]
      summary: Sincronizacion batch de mediciones wearable
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncMeasurementsRequest'
      responses:
        '200':
          description: Mediciones sincronizadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncMeasurementsResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/subscriptions/webhook:
    post:
      tags: [SUBSCRIPTIONS]
      summary: Webhook de Stripe validado por firma
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Evento aceptado
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [success] }
        '400': { $ref: '#/components/responses/Error' }
  /v1/admin/break-glass/request:
    post:
      tags: [ADMIN]
      summary: Solicitar acceso break-glass de emergencia
      description: Solo admin. Inicia el flujo Two-Person rule; otro admin debe aprobarlo. Duracion maxima 2 horas.
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BreakGlassRequestInput'
      responses:
        '201':
          description: Solicitud break-glass pendiente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreakGlassRequest'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/admin/break-glass/{id}/approve:
    post:
      tags: [ADMIN]
      summary: Aprobar acceso break-glass
      description: Solo admin. El aprobador debe ser distinto al solicitante.
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: bgr_123 }
      responses:
        '200':
          description: Solicitud break-glass aprobada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BreakGlassRequest'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /v1/alerts/{id}:
    get:
      tags: [ALERTAS]
      summary: Detalle de alerta
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: alrt_445566 }
        - name: X-CSRF-Token
          in: header
          required: false
          description: Required when authenticating via HttpOnly cookie.
          schema: { type: string }
      responses:
        '200':
          description: Alerta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/alerts/{id}/acknowledge:
    post:
      tags: [ALERTAS]
      summary: Reconocer una alerta generada por anomalía
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: alrt_445566 }
        - name: X-CSRF-Token
          in: header
          required: false
          description: Required when authenticating via HttpOnly cookie.
          schema: { type: string }
      responses:
        '200':
          description: Alerta reconocida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/patients/{id}:
    get:
      tags: [PACIENTES]
      summary: Obtener perfil del paciente para cuidador o web
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: usr_67a1b2c3d4e5 }
        - name: X-CSRF-Token
          in: header
          required: false
          description: Required when authenticating via HttpOnly cookie.
          schema: { type: string }
      responses:
        '200':
          description: Perfil del paciente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientProfile'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /v1/patients/{id}/measurements:
    get:
      tags: [SYNC]
      summary: Listar mediciones biométricas time-series del paciente
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientID'
        - name: type
          in: query
          required: false
          schema: { type: string, enum: [heart_rate, blood_oxygen, steps] }
        - name: from
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
      responses:
        '200':
          description: Mediciones ordenadas por timestamp descendente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementListResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/patients/{id}/clinical-records:
    get:
      tags: [CLINICAL]
      summary: Listar registros clinicos del paciente
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientID'
      responses:
        '200':
          description: Registros clinicos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicalRecordListResponse'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [CLINICAL]
      summary: Crear registro clinico del paciente
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClinicalRecordRequest'
      responses:
        '201':
          description: Registro clinico creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClinicalRecordResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/patients/{id}/medications:
    get:
      tags: [CLINICAL]
      summary: Listar medicamentos activos o historicos del paciente
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientID'
      responses:
        '200':
          description: Medicamentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationListResponse'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [CLINICAL]
      summary: Crear medicamento del pastillero
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMedicationRequest'
      responses:
        '201':
          description: Medicamento creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/patients/{id}/medication-logs:
    post:
      tags: [CLINICAL]
      summary: Registrar toma o evento puntual de medicacion
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMedicationLogRequest'
      responses:
        '201':
          description: Log de medicacion creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationLogResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/patients/{id}/reports:
    get:
      tags: [REPORTS]
      summary: Listar referencias a reportes PDF del paciente en S3
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientID'
      responses:
        '200':
          description: Reportes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportListResponse'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [REPORTS]
      summary: Registrar referencia a reporte PDF generado
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PatientID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201':
          description: Reporte registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/consents:
    post:
      tags: [CONSENT]
      summary: Otorgar scopes granulares de paciente a cuidador
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantConsentRequest'
      responses:
        '201':
          description: Consentimiento creado o actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/consents/{caregiver_id}:
    delete:
      tags: [CONSENT]
      summary: Revocar todos los scopes otorgados a un cuidador
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CaregiverID'
      responses:
        '200':
          description: Consentimiento revocado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/relationships:
    get:
      tags: [RELATIONSHIPS]
      summary: Listar relaciones caregiver-paciente visibles al usuario autenticado
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Relaciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipListResponse'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [RELATIONSHIPS]
      summary: Asignar un cuidador al paciente autenticado
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRelationshipRequest'
      responses:
        '201':
          description: Relacion activa creada o reactivada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/relationships/{caregiver_id}:
    delete:
      tags: [RELATIONSHIPS]
      summary: Revocar relacion activa entre paciente autenticado y cuidador
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CaregiverID'
      responses:
        '200':
          description: Relacion revocada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/devices:
    get:
      tags: [DEVICES]
      summary: Listar wearables vinculados al usuario autenticado
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Dispositivos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      tags: [DEVICES]
      summary: Registrar wearable vinculado al usuario autenticado
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: Dispositivo registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/devices/{id}/transfer-requests:
    post:
      tags: [DEVICES]
      summary: Solicitar transferencia de propiedad de wearable
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: dev_998877 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTransferRequestInput'
      responses:
        '201':
          description: Solicitud de transferencia creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTransferResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/device-transfer-requests/{id}/approve:
    post:
      tags: [DEVICES]
      summary: Aprobar transferencia pendiente de wearable
      description: Solo el usuario destino de la transferencia puede aprobarla; al aprobar cambia el propietario del dispositivo.
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: dtr_123 }
      responses:
        '200':
          description: Transferencia aprobada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTransferResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/device-transfer-requests/{id}/reject:
    post:
      tags: [DEVICES]
      summary: Rechazar transferencia pendiente de wearable
      description: El usuario origen o destino de la transferencia puede rechazarla.
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: dtr_123 }
      responses:
        '200':
          description: Transferencia rechazada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTransferResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  parameters:
    PatientID:
      name: id
      in: path
      required: true
      schema: { type: string, example: usr_67a1b2c3d4e5 }
...
  responses:
    Error:
      description: Error JSON
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, role, first_name, last_name]
      additionalProperties: false
      properties:
        email: { type: string, format: email, example: juan@example.com }
        password: { type: string, minLength: 9, example: Secure!1234 }
        role: { type: string, enum: [patient, caregiver], example: patient }
        first_name: { type: string, example: Juan }
        last_name: { type: string, example: Perez }
    RegisterResponse:
      type: object
      required: [status, data]
      properties:
        status: { type: string, enum: [success] }
        data:
          type: object
          required: [user_id, message]
          properties:
            user_id: { type: string, example: usr_67a1b2c3d4e5 }
            message: { type: string, example: User registered successfully. Please verify your email. }
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, format: email, example: juan@example.com }
        password: { type: string, example: Secure!1234 }
...
    MobileLoginResponse:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer, example: 900 }
...